+++ packages/swirl-tokens/scripts/build.ts
@@ -23,16 +23,7 @@ StyleDictionary.registerFormat({
           mappedTokens[mappedType as keyof mappedTokensType] = {};
         }
 
-        if (theme === "light") {
-          mappedTokens[mappedType as keyof mappedTokensType][name] = {
-            DEFAULT: value,
-            [`${theme}`]: value,
-          };
-        } else {
-          mappedTokens[mappedType as keyof mappedTokensType][name] = {
-            [`${theme}`]: value,
-          };
-        }
+        mappedTokens[mappedType as keyof mappedTokensType][name] = value;
       }
     });
 
@@ -134,6 +125,35 @@ StyleDictionary.registerTransform({
     };
 
     return Object.assign(generatedAttrs, originalAttrs);
+  },
+});
+
+StyleDictionary.registerTransform({
+  name: "shadow/css",
+  type: "value",
+  transitive: true,
+  matcher: (token) => token.type === "boxShadow",
+  transformer: function (token) {
+    const shadows = Array.isArray(token.value) ? token.value : [token.value];
+
+    const transformedShadows = shadows.map((shadow) => {
+      const { x, y, blur, spread, color, type } = shadow;
+      const inset = type === "innerShadow" ? "inset " : "";
+
+      return `${inset}${x}px ${y}px ${blur}px ${spread}px ${color}`;
+    });
+
+    return transformedShadows.join(", ");
+  },
+});
+
+StyleDictionary.registerTransform({
+  name: "fontWeight/flutter",
+  type: "value",
+  transitive: true,
+  matcher: (token) => token.type === "fontWeights",
+  transformer: function (token) {
+    return token.value.replace(/^"(\d+)"$/, "FontWeight.w$1");
   },
 });