+++ packages/swirl-components/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @getflip/swirl-components
 
+## 0.9.1
+
+### Patch Changes
+
+- Updated dependencies
+  [[`3d28b36`](https://github.com/getflip/swirl/commit/3d28b36463b91c6012b8bd075e69cd7493e5b13d)]:
+  - @getflip/swirl-tokens@2.2.0
+
+## 0.9.0
+
+### Minor Changes
+
+- [#97](https://github.com/getflip/swirl/pull/97)
+  [`a241f77`](https://github.com/getflip/swirl/commit/a241f778d415d44cf116af0cf671c278384e1c2b)
+  Thanks [@Sqrrl](https://github.com/Sqrrl)! - Introduce shadow design tokens
+
+### Patch Changes
+
+- [#108](https://github.com/getflip/swirl/pull/108)
+  [`1fcca44`](https://github.com/getflip/swirl/commit/1fcca448df5bf8f2cbc03d12f6feccaa830b01f5)
+  Thanks [@Sqrrl](https://github.com/Sqrrl)! - Tweak component focus styles
+
+- Updated dependencies
+  [[`a241f77`](https://github.com/getflip/swirl/commit/a241f778d415d44cf116af0cf671c278384e1c2b)]:
+  - @getflip/swirl-tokens@2.1.0
+
 ## 0.8.6
 
 ### Patch Changes